+++ src/components/locales/i18n.js
@@ -1,42 +1,42 @@
-import i18next from 'i18next';
-import { initReactI18next } from 'react-i18next';
-import LanguageDetector from 'i18next-browser-languagedetector';
+import i18next from "i18next";
+import { initReactI18next } from "react-i18next";
+import LanguageDetector from "i18next-browser-languagedetector";
 
 i18next
   .use(LanguageDetector)
   .use(initReactI18next)
   .init({
     debug: true,
-    fallbackLng: 'tr',
+    fallbackLng: "tr",
     interpolation: {
       escapeValue: false,
     },
     resources: {
       en: {
         translation: {
-          "ANA SAYFA": "HOME",
-          "HAKKIMIZDA": "ABOUT",
-          "PARTNERLERİMİZ": "PARTNERS",
-          "İLETİŞİM": "CONTACT"
-        }
+          home: "HOME",
+          about: "ABOUT",
+          partners: "PARTNERS",
+          contact: "CONTACT",
+        },
       },
       de: {
         translation: {
-          "ANA SAYFA": "STARTSEITE",
-          "HAKKIMIZDA": "ÜBER UNS",
-          "PARTNERLERİMİZ": "PARTNER",
-          "İLETİŞİM": "KONTAKT"
-        }
+          home: "STARTSEITE",
+          about: "ÜBER UNS",
+          partners: "PARTNER",
+          contact: "KONTAKT",
+        },
       },
       tr: {
         translation: {
-          "ANA SAYFA": "ANA SAYFA",
-          "HAKKIMIZDA": "HAKKIMIZDA",
-          "PARTNERLERİMİZ": "PARTNERLERİMİZ",
-          "İLETİŞİM": "İLETİŞİM"
-        }
-      }
-    }
+          home: "ANA SAYFA",
+          about: "HAKKIMIZDA",
+          partners: "PARTNERLERİMİZ",
+          contact: "İLETİŞİM",
+        },
+      },
+    },
   });
 
 export default i18next;