+++ src/main.tsx
@@ -1,7 +1,6 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App.tsx";
-import './components/locales/i18n';
 import "bootstrap/dist/css/bootstrap.css";
 
 createRoot(document.getElementById("root")!).render(