+++ src/pages/Contact/cards.tsx
@@ -1,14 +1,16 @@
 import { Row, Col, Card } from "react-bootstrap";
 import { GeoAltFill, TelephoneFill, EnvelopeFill } from "react-bootstrap-icons";
+import { useTranslation } from "react-i18next";
 
 const Cards = () => {
+  const { t } = useTranslation();
   return (
     <Row className="mb-5 mt-3 d-flex align-items-stretch">
       <Col md={4} className="d-flex">
         <Card className="text-center h-100 w-100">
           <Card.Body>
             <GeoAltFill className="text-primary mb-3" size={40} />
-            <Card.Title>Ofisimiz</Card.Title>
+            <Card.Title>{t("office")}</Card.Title>
             <Card.Text>
               NEF 22 A Blok. Ataköy 7-8-9-10. Kısım Mh. Çobançeşme E-5 Yan Yol
               Cad. D:30 Bakırköy/İstanbul
@@ -20,9 +22,9 @@ const Cards = () => {
         <Card className="text-center h-100 w-100">
           <Card.Body>
             <TelephoneFill className="text-primary mb-3" size={40} />
-            <Card.Title>Mobil</Card.Title>
+            <Card.Title>{t("mobile")}</Card.Title>
             <Card.Text>+90 (542) 194 14 13</Card.Text>
-            <Card.Title>Telefon</Card.Title>
+            <Card.Title>{t("phone")}</Card.Title>
             <Card.Text>+90 (212) 936 19 61</Card.Text>
           </Card.Body>
         </Card>
@@ -31,7 +33,7 @@ const Cards = () => {
         <Card className="text-center h-100 w-100">
           <Card.Body>
             <EnvelopeFill className="text-primary mb-3" size={40} />
-            <Card.Title>E-Posta</Card.Title>
+            <Card.Title>{t("email")}</Card.Title>
             <Card.Text>
               <a
                 href="mailto:kamuran.altay@fu-ra.com.tr"