+++ src/pages/home/Home.tsx
@@ -1,0 +1,17 @@
+import Carousel from "./carousel";
+import Section from "./section";
+import Footer from "../../components/Footer";
+import PartnersChild from "../partners/PartnersChild";
+
+const Home = () => {
+  return (
+    <>
+      <Carousel />
+      <Section />
+      <PartnersChild />
+      <Footer />
+    </>
+  );
+};
+
+export default Home;